refactor(subscriptions): condense SubscriptionModule + move feature-plan resolver
Per review feedback:
- Condense canQuerySubscriptionPermission, shouldCreateNativeFreeSubscription
  with `||` instead of separate early-exit ifs.
- Convert canAddDatasets, canInviteMembers, grantsWorkspaceEntitlements
  to positive-assertion form (`if (subscription) { … } return false`).
- Document why buildNativeFreeFieldsForDB lives on SubscriptionModule
  rather than SubscriptionParsers (Deno-compatibility for edge functions).
- Move resolveFeaturePlanTypeForWorkspace out of src/hooks (it isn't a
  hook) into SubscriptionModule.resolveFeaturePlanTypeForWorkspace, with
  ResolvedFeaturePlanType promoted to Subscription.types.
- Move the resolver's tests into SubscriptionModule.test.ts.

Addresses review threads at SubscriptionModule.ts L48, L138, L167,
L239-246; resolveFeaturePlanTypeForWorkspace.ts L14.
